requirements.txt
pandas
numpy
openpyxl
scikit-learn
ortools
requests
pyyaml
folium
tqdm
